Rename/addgame to addgamescreen (the class name)
