+++ lib/screens/home.dart
@@ -2,7 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:gamelib_mob/list/list_class.dart';
 import 'package:gamelib_mob/list/main_list.dart';
-import 'package:gamelib_mob/screens/add_game.dart';
+import 'package:gamelib_mob/screens/add_game_screen.dart';
 import 'package:gamelib_mob/screens/game_page.dart';
 import 'package:gamelib_mob/screens/sign_in.dart';
 import 'package:gamelib_mob/helpers/helpers.dart';
